Move Sorting to Search namespace
