+++ app/Console/Command/ConvertDescriptionToEducationalDescriptionForCultuurkuur.php
@@ -12,7 +12,7 @@ use CultuurNet\UDB3\ReadModel\DocumentDoesNotExist;
 use CultuurNet\UDB3\ReadModel\DocumentRepository;
 use CultuurNet\UDB3\Search\OrganizersSapi3SearchService;
 use CultuurNet\UDB3\Search\ResultsGenerator;
-use CultuurNet\UDB3\Sorting;
+use CultuurNet\UDB3\Search\Sorting;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
