+++ app/Console/Command/UpdateEventsAttendanceMode.php
@@ -11,7 +11,7 @@ use CultuurNet\UDB3\Event\ValueObjects\LocationId;
 use CultuurNet\UDB3\Model\ValueObject\Online\AttendanceMode;
 use CultuurNet\UDB3\Search\ResultsGenerator;
 use CultuurNet\UDB3\Search\SearchServiceInterface;
-use CultuurNet\UDB3\Sorting;
+use CultuurNet\UDB3\Search\Sorting;
 use Exception;
 use RuntimeException;
 use Symfony\Component\Console\Helper\ProgressBar;
