+++ src/EventExport/EventExportServiceInterface.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\EventExport;
 
 use CultuurNet\UDB3\Model\ValueObject\Web\EmailAddress;
-use CultuurNet\UDB3\Sorting;
+use CultuurNet\UDB3\Search\Sorting;
 use Psr\Log\LoggerInterface;
 
 interface EventExportServiceInterface
