+++ src/Search/ResultsGeneratorInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Search;
 
-use CultuurNet\UDB3\Sorting;
 use Generator;
 
 interface ResultsGeneratorInterface
