+++ src/Search/Sorting.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3;
+namespace CultuurNet\UDB3\Search;
 
 use CultuurNet\UDB3\Deserializer\MissingValueException;
